refactor(workflow): hoist StagedApprovalError to the shared seam
It is consumed by both the MCP registry and the workflow engine. Hoisting
it keeps errors.As matching across the coming module split; a per-module
twin would silently degrade to the bare sentinel and lose the approval id.
